problem-hunt/python-function/requirements.txt
azure-functions==1.20.0
azure-cosmos==4.5.1
PyJWT==2.9.0
requests==2.31.0
supabase==2.3.4
